fixed some standard errors
