+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+  "solidity.compileUsingRemoteVersion": "v0.6.10+commit.00c0fcaf"
+}
