--- packages/buidler/artifacts/Address.json
@@ -1,8 +1,0 @@
-{
-  "contractName": "Address",
-  "abi": [],
-  "bytecode": "0x60566023600b82828239805160001a607314601657fe5b30600052607381538281f3fe73000000000000000000000000000000000000000030146080604052600080fdfea26469706673582212205c5f66509de853397494cfd47e53507f72a718ddac7e6a3dbfd379a0cc228c6164736f6c63430006060033",
-  "deployedBytecode": "0x73000000000000000000000000000000000000000030146080604052600080fdfea26469706673582212205c5f66509de853397494cfd47e53507f72a718ddac7e6a3dbfd379a0cc228c6164736f6c63430006060033",
-  "linkReferences": {},
-  "deployedLinkReferences": {}
-}
