--- packages/buidler/artifacts/Counters.json
@@ -1,8 +1,0 @@
-{
-  "contractName": "Counters",
-  "abi": [],
-  "bytecode": "0x60566023600b82828239805160001a607314601657fe5b30600052607381538281f3fe73000000000000000000000000000000000000000030146080604052600080fdfea26469706673582212208f622ff925f3c0fc6d56564e4815e63334f297e8628b589faba07734821b2a9c64736f6c63430006060033",
-  "deployedBytecode": "0x73000000000000000000000000000000000000000030146080604052600080fdfea26469706673582212208f622ff925f3c0fc6d56564e4815e63334f297e8628b589faba07734821b2a9c64736f6c63430006060033",
-  "linkReferences": {},
-  "deployedLinkReferences": {}
-}
