--- packages/buidler/artifacts/EnumerableMap.json
@@ -1,8 +1,0 @@
-{
-  "contractName": "EnumerableMap",
-  "abi": [],
-  "bytecode": "0x60566023600b82828239805160001a607314601657fe5b30600052607381538281f3fe73000000000000000000000000000000000000000030146080604052600080fdfea264697066735822122091366248673e4b1be8ee609c8c191bf68d0d14219de696f832305741f59b939e64736f6c63430006060033",
-  "deployedBytecode": "0x73000000000000000000000000000000000000000030146080604052600080fdfea264697066735822122091366248673e4b1be8ee609c8c191bf68d0d14219de696f832305741f59b939e64736f6c63430006060033",
-  "linkReferences": {},
-  "deployedLinkReferences": {}
-}
