--- packages/buidler/artifacts/EnumerableSet.json
@@ -1,8 +1,0 @@
-{
-  "contractName": "EnumerableSet",
-  "abi": [],
-  "bytecode": "0x60566023600b82828239805160001a607314601657fe5b30600052607381538281f3fe73000000000000000000000000000000000000000030146080604052600080fdfea264697066735822122007e7d10e99692d066b91735af82a5da31d35d5f41859e77c9b17bfdd72a18f0864736f6c63430006060033",
-  "deployedBytecode": "0x73000000000000000000000000000000000000000030146080604052600080fdfea264697066735822122007e7d10e99692d066b91735af82a5da31d35d5f41859e77c9b17bfdd72a18f0864736f6c63430006060033",
-  "linkReferences": {},
-  "deployedLinkReferences": {}
-}
