--- packages/buidler/artifacts/Game.json
@@ -1,30 +1,0 @@
-{
-  "contractName": "Game",
-  "abi": [
-    {
-      "inputs": [
-        {
-          "internalType": "string",
-          "name": "_gameName",
-          "type": "string"
-        },
-        {
-          "internalType": "uint256",
-          "name": "_height",
-          "type": "uint256"
-        },
-        {
-          "internalType": "uint256",
-          "name": "_width",
-          "type": "uint256"
-        }
-      ],
-      "stateMutability": "nonpayable",
-      "type": "constructor"
-    }
-  ],
-  "bytecode": "0x608060405234801561001057600080fd5b506040516101fb3803806101fb8339818101604052606081101561003357600080fd5b810190808051604051939291908464010000000082111561005357600080fd5b90830190602082018581111561006857600080fd5b825164010000000081118282018810171561008257600080fd5b82525081516020918201929091019080838360005b838110156100af578181015183820152602001610097565b50505050905090810190601f1680156100dc5780820380516001836020036101000a031916815260200191505b50604090815260208281015192909101518551929450925061010391600091860190610113565b50600191909155600255506101ae565b828054600181600116156101000203166002900490600052602060002090601f016020900481019282601f1061015457805160ff1916838001178555610181565b82800160010185558215610181579182015b82811115610181578251825591602001919060010190610166565b5061018d929150610191565b5090565b6101ab91905b8082111561018d5760008155600101610197565b90565b603f806101bc6000396000f3fe6080604052600080fdfea2646970667358221220c19748afd707efda23ef3258cbeb8549633ed935b828c973ceb59a12d71189b064736f6c63430006060033",
-  "deployedBytecode": "0x6080604052600080fdfea2646970667358221220c19748afd707efda23ef3258cbeb8549633ed935b828c973ceb59a12d71189b064736f6c63430006060033",
-  "linkReferences": {},
-  "deployedLinkReferences": {}
-}
