--- packages/buidler/artifacts/SafeMath.json
@@ -1,8 +1,0 @@
-{
-  "contractName": "SafeMath",
-  "abi": [],
-  "bytecode": "0x60566023600b82828239805160001a607314601657fe5b30600052607381538281f3fe73000000000000000000000000000000000000000030146080604052600080fdfea264697066735822122041bfa5a40fe1564a083729f3daaea15577864393e7818302eb9ed2696aed937364736f6c63430006060033",
-  "deployedBytecode": "0x73000000000000000000000000000000000000000030146080604052600080fdfea264697066735822122041bfa5a40fe1564a083729f3daaea15577864393e7818302eb9ed2696aed937364736f6c63430006060033",
-  "linkReferences": {},
-  "deployedLinkReferences": {}
-}
