--- packages/buidler/artifacts/console.json
@@ -1,8 +1,0 @@
-{
-  "contractName": "console",
-  "abi": [],
-  "bytecode": "0x60566023600b82828239805160001a607314601657fe5b30600052607381538281f3fe73000000000000000000000000000000000000000030146080604052600080fdfea2646970667358221220d8aaa6f0d3d64b2d903e6312055543f062ccf7fd4416c4883926bae0918a975b64736f6c63430006060033",
-  "deployedBytecode": "0x73000000000000000000000000000000000000000030146080604052600080fdfea2646970667358221220d8aaa6f0d3d64b2d903e6312055543f062ccf7fd4416c4883926bae0918a975b64736f6c63430006060033",
-  "linkReferences": {},
-  "deployedLinkReferences": {}
-}
