--- packages/buidler/cache/console-history.txt
@@ -1,5 +1,0 @@
-exit
-ls
-ethereum
-web3.
-var game = new Game()
