--- packages/buidler/cache/last-solc-config.json
@@ -1,10 +1,0 @@
-{
-  "solc": {
-    "version": "0.6.6",
-    "optimizer": {
-      "enabled": true,
-      "runs": 200
-    }
-  },
-  "buidlerVersion": "1.3.8"
-}
