+++ packages/client/.gitignore
@@ -8,6 +8,9 @@
 # testing
 /coverage
 
+# contract builds
+/contracts
+
 # production
 /build
