+++ packages/client/src/pages/planet/tests/bitmap.js
@@ -1,0 +1,2 @@
+module.exports =
+  '00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000011111100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000011111111101111000000000000000000000011000000000000000000000000000000000000000000000000000000000000000000001111011111111111111110000000000000011000000111000000000111000000000000000000000000000000000000000000000000000011001111111111111111111100000000000111100000000111000000010000000000000000000000000000000000000000000100000000001111111111111111111111110000001100000000000000001111000000010100000011111100000000000000000000000111110000000011011011111101111111111110000001111000000000000001111110011000111111111111111000000000000000000000011111110000011101111111100011111111111000000111000000001100001111111111111111111111111111100000000000000000000111111111111110111101101110000011111111100000000000000001000001111111111111111111111111111100000000000000000001011111111111111111110111111000001111111110000000000000000100011111111111111111111111111111111000000000000000000001111111111111110011111011110001111111110000000000000000010001111111111111111111111111111111000000000000000000001111111111111111111111111111000111111111000000000010000000111111111111111111111111111111111000000000000000000000111111111111111111111111011100001111111000000000011111001111111111111111111111111111111011000000000000000000000001111111111111111111111101111101111100000000000011111110111111111111111111111111111111101100000000000000000000001110001111111111111110111011000111100001100000001111110111111111111111111111111111110000111000000000000000000011000000011111111111100001111100011100001110000001111111111111111111111111111111111111000011100000000000000110000000000001111111111110000011110000110000000000001110111111111111111111111111111111111100000110000000000000000000000000000111111111111100001111000000000000000000111101111111111111111111111111111111111000001000000000000000000000000000001111111111111101111100000000000000111001111111111111111111111111111111111111110000000000000000000000000000000000111111111111111111111000000000000011100111111111111111111111111111111111111111100000000000000000000000000000000011111111111111111111100000000000001111111111111111111111111111111111111111111110010000000000000000000000000000001111111111111111110111000000000000011111111111111111111111111111111111111111100010000000000000000000000000000000111111111111111111100000000000000000111111111111111111111111111111111111111100011000000000000000000000000000000011111111111111110000000000000000001111101111100011111111111111111111111111111001100000000000000000000000000000001111111111111110000000000000000000111001011111111111111111111111111111111101101100000000000000000000000000000000111111111111110000000000000000000011100011011111111111111111111111111111100101100000000000000000000000000000000001111111111110000000000000000000000111110000101111111111111111111111111111001000000000000000000000000000000000000111111111110000000000000000000000111111111111111111111111111111111111111110000000000000000000000000000000000000001111100001000000000000000000001111111111111111111111111111111111111111110010000000000000000000000000000000000000111110000000000000000000000000111111111111111111111001111111111111111110100000000000000000000000000000000000000000111001011110000000000000000011111111111111110111111000111111111111110000000000000000000000000000000000000000000011111100010110000000000000001111111111111111111111100000111000111100001000000000000000000000000000000000000000000000111000000000000000000001111111111111111110111000000011000001111000100000000000000000000000000000000000000000000000110111110000000000000011111111111111111111100000000100000001100001100000000000000000000000000000000000000000000001111111100000000000000111111111111111111110000000001000101000010011000000000000000000000000000000000000000000000011111111100000000000000000011111111111110000000000000001100011101000000000000000000000000000000000000000000000001111111111000000000000000000111111111100000000000000000011011111010000000000000000000000000000000000000000000001111111111111100000000000000011111111110000000000000000001110011111011101100000000000000000000000000000000000000111111111111111000000000000001111111110000000000000000000001100110000111111000000000000000000000000000000000000001111111111111100000000000000011111111000000000000000000000000001100000011011000000000000000000000000000000000000011111111111100000000000000011111111100010000000000000000000000000110100000000000000000000000000000000000000000001111111111110000000000000001111111110011000000000000000000000001111011000000100000000000000000000000000000000000001111111111000000000000000011111110011100000000000000000000001111111100000000010000000000000000000000000000000000111111111000000000000000001111111001100000000000000000000011111111111000011000000000000000000000000000000000000011111110000000000000000000111111000000000000000000000000011111111111100000000000000000000000000000000000000000001111111000000000000000000001111100000000000000000000000001111111111111000000000000000000000000000000000000000000111111000000000000000000000111000000000000000000000000000111111111111000000000000000000000000000000000000000000011111000000000000000000000000000000000000000000000000000001100011111100000000000000000000000000000000000000000001111000000000000000000000000000000000000000000000000000000000001111100000000000000000000000000000000000000000000111000000000000000000000000000000000000000000000000000000000000011110000000100000000000000000000000000000000000011100000000000000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000001110000000000000000000000000000000000000000000000000000000000000011000000110000000000000000000000000000000000000111000000000000000000000000000000000000000000000000000000000000000000000110000000000000000000000000000000000000011101000000000000000000000000000000000000000000000000000000000000000000110000000000000000000000000000000000000000111000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000'
