+++ packages/client/src/pages/planet/tests/map.js
@@ -1,0 +1,898 @@
+module.exports = [
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  31,
+  128,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  63,
+  239,
+  0,
+  0,
+  3,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  15,
+  127,
+  255,
+  128,
+  1,
+  129,
+  192,
+  28,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  51,
+  255,
+  255,
+  192,
+  7,
+  128,
+  112,
+  16,
+  0,
+  0,
+  0,
+  0,
+  1,
+  0,
+  63,
+  255,
+  255,
+  192,
+  192,
+  0,
+  60,
+  5,
+  3,
+  240,
+  0,
+  0,
+  31,
+  0,
+  219,
+  247,
+  255,
+  129,
+  224,
+  0,
+  126,
+  99,
+  255,
+  248,
+  0,
+  0,
+  31,
+  193,
+  223,
+  227,
+  255,
+  129,
+  192,
+  48,
+  255,
+  255,
+  255,
+  248,
+  0,
+  0,
+  127,
+  254,
+  246,
+  224,
+  255,
+  128,
+  0,
+  65,
+  255,
+  255,
+  255,
+  240,
+  0,
+  1,
+  127,
+  255,
+  247,
+  224,
+  255,
+  128,
+  0,
+  71,
+  255,
+  255,
+  255,
+  248,
+  0,
+  0,
+  127,
+  255,
+  62,
+  241,
+  255,
+  0,
+  0,
+  71,
+  255,
+  255,
+  255,
+  240,
+  0,
+  0,
+  255,
+  255,
+  255,
+  241,
+  255,
+  0,
+  32,
+  63,
+  255,
+  255,
+  255,
+  224,
+  0,
+  0,
+  255,
+  255,
+  255,
+  112,
+  254,
+  0,
+  124,
+  255,
+  255,
+  255,
+  254,
+  192,
+  0,
+  0,
+  127,
+  255,
+  255,
+  125,
+  240,
+  0,
+  254,
+  255,
+  255,
+  255,
+  254,
+  192,
+  0,
+  0,
+  227,
+  255,
+  251,
+  177,
+  225,
+  128,
+  253,
+  255,
+  255,
+  255,
+  240,
+  224,
+  0,
+  3,
+  1,
+  255,
+  225,
+  241,
+  195,
+  129,
+  255,
+  255,
+  255,
+  255,
+  240,
+  224,
+  0,
+  96,
+  1,
+  255,
+  224,
+  240,
+  192,
+  3,
+  191,
+  255,
+  255,
+  255,
+  240,
+  96,
+  0,
+  0,
+  1,
+  255,
+  240,
+  240,
+  0,
+  3,
+  223,
+  255,
+  255,
+  255,
+  248,
+  32,
+  0,
+  0,
+  0,
+  255,
+  253,
+  240,
+  0,
+  57,
+  255,
+  255,
+  255,
+  255,
+  252,
+  0,
+  0,
+  0,
+  0,
+  255,
+  255,
+  248,
+  0,
+  57,
+  255,
+  255,
+  255,
+  255,
+  254,
+  0,
+  0,
+  0,
+  0,
+  255,
+  255,
+  248,
+  0,
+  63,
+  255,
+  255,
+  255,
+  255,
+  254,
+  64,
+  0,
+  0,
+  0,
+  255,
+  255,
+  220,
+  0,
+  31,
+  255,
+  255,
+  255,
+  255,
+  248,
+  128,
+  0,
+  0,
+  0,
+  255,
+  255,
+  224,
+  0,
+  15,
+  255,
+  255,
+  255,
+  255,
+  241,
+  128,
+  0,
+  0,
+  0,
+  255,
+  255,
+  0,
+  0,
+  62,
+  248,
+  255,
+  255,
+  255,
+  249,
+  128,
+  0,
+  0,
+  0,
+  255,
+  254,
+  0,
+  0,
+  57,
+  127,
+  255,
+  255,
+  255,
+  219,
+  0,
+  0,
+  0,
+  0,
+  255,
+  252,
+  0,
+  0,
+  56,
+  223,
+  255,
+  255,
+  255,
+  150,
+  0,
+  0,
+  0,
+  0,
+  127,
+  248,
+  0,
+  0,
+  31,
+  11,
+  255,
+  255,
+  255,
+  200,
+  0,
+  0,
+  0,
+  0,
+  127,
+  240,
+  0,
+  0,
+  63,
+  255,
+  255,
+  255,
+  255,
+  224,
+  0,
+  0,
+  0,
+  0,
+  62,
+  16,
+  0,
+  0,
+  255,
+  255,
+  255,
+  255,
+  255,
+  200,
+  0,
+  0,
+  0,
+  0,
+  62,
+  0,
+  0,
+  0,
+  255,
+  255,
+  249,
+  255,
+  255,
+  160,
+  0,
+  0,
+  0,
+  0,
+  14,
+  94,
+  0,
+  0,
+  255,
+  255,
+  126,
+  63,
+  255,
+  0,
+  0,
+  0,
+  0,
+  0,
+  15,
+  197,
+  128,
+  0,
+  255,
+  255,
+  254,
+  14,
+  60,
+  32,
+  0,
+  0,
+  0,
+  0,
+  0,
+  224,
+  0,
+  1,
+  255,
+  255,
+  184,
+  12,
+  30,
+  32,
+  0,
+  0,
+  0,
+  0,
+  0,
+  55,
+  192,
+  0,
+  255,
+  255,
+  248,
+  4,
+  6,
+  24,
+  0,
+  0,
+  0,
+  0,
+  0,
+  31,
+  224,
+  0,
+  127,
+  255,
+  248,
+  2,
+  40,
+  76,
+  0,
+  0,
+  0,
+  0,
+  0,
+  15,
+  248,
+  0,
+  1,
+  255,
+  240,
+  0,
+  24,
+  232,
+  0,
+  0,
+  0,
+  0,
+  0,
+  15,
+  252,
+  0,
+  0,
+  255,
+  192,
+  0,
+  13,
+  244,
+  0,
+  0,
+  0,
+  0,
+  0,
+  31,
+  255,
+  128,
+  0,
+  255,
+  192,
+  0,
+  14,
+  125,
+  216,
+  0,
+  0,
+  0,
+  0,
+  31,
+  255,
+  192,
+  0,
+  255,
+  128,
+  0,
+  3,
+  48,
+  252,
+  0,
+  0,
+  0,
+  0,
+  15,
+  255,
+  192,
+  0,
+  127,
+  128,
+  0,
+  0,
+  24,
+  27,
+  0,
+  0,
+  0,
+  0,
+  7,
+  255,
+  128,
+  0,
+  255,
+  136,
+  0,
+  0,
+  3,
+  64,
+  0,
+  0,
+  0,
+  0,
+  7,
+  255,
+  128,
+  0,
+  255,
+  152,
+  0,
+  0,
+  15,
+  96,
+  64,
+  0,
+  0,
+  0,
+  1,
+  255,
+  128,
+  0,
+  127,
+  56,
+  0,
+  0,
+  31,
+  224,
+  8,
+  0,
+  0,
+  0,
+  1,
+  255,
+  0,
+  0,
+  127,
+  48,
+  0,
+  0,
+  127,
+  240,
+  192,
+  0,
+  0,
+  0,
+  1,
+  252,
+  0,
+  0,
+  126,
+  0,
+  0,
+  0,
+  255,
+  240,
+  0,
+  0,
+  0,
+  0,
+  1,
+  252,
+  0,
+  0,
+  62,
+  0,
+  0,
+  0,
+  255,
+  248,
+  0,
+  0,
+  0,
+  0,
+  1,
+  248,
+  0,
+  0,
+  56,
+  0,
+  0,
+  0,
+  255,
+  240,
+  0,
+  0,
+  0,
+  0,
+  1,
+  240,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  99,
+  240,
+  0,
+  0,
+  0,
+  0,
+  1,
+  224,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  3,
+  224,
+  0,
+  0,
+  0,
+  0,
+  1,
+  192,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  1,
+  224,
+  32,
+  0,
+  0,
+  0,
+  1,
+  192,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  32,
+  0,
+  0,
+  0,
+  1,
+  192,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  192,
+  192,
+  0,
+  0,
+  0,
+  1,
+  192,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  1,
+  128,
+  0,
+  0,
+  0,
+  1,
+  208,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  3,
+  0,
+  0,
+  0,
+  0,
+  0,
+  224,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0
+]
